Add files via upload
The h images are some samples of a healthcare dashboard, this dashboard was implemented for an enterprise in my last job, there are different kind of graphs, dynamic filters, buttons to navigate as well as markers, tooltips to detailed some insights. Also a model was developed in order to manage data in a suitable way, some columns, transformations and measurements were added to normalized data and create better visuals.